+++ packages/build-app-helpers/tests/fixtures/analyzeAuth/app.ts
@@ -1,0 +1,7 @@
+import { type AppConfig, runApp } from 'ice';
+
+const appConfig: AppConfig = {
+  auth: {},
+}
+
+runApp(appConfig);