+++ packages/vite-service/src/plugins/history.ts
@@ -1,5 +1,5 @@
 import type { Plugin } from 'vite';
-import { isBoolean } from 'lodash';
+import { isBoolean, isUndefined } from 'lodash';
 import { all } from 'deepmerge';
 import * as history from 'connect-history-api-fallback';
 
@@ -9,7 +9,7 @@ import * as history from 'connect-history-api-fallback';
  * @see https://github.com/vitejs/vite/blob/8733a83d291677b9aff9d7d78797ebb44196596e/packages/vite/src/node/server/index.ts#L433
  */
 export const serverHistoryPlugin = (opt: history.Options | boolean): Plugin => {
-  if (isBoolean(opt)) return null;
+  if (isBoolean(opt) || isUndefined(opt)) return null;
 
   return {
     name: 'vite-plugin-server-history',