+++ packages/build-app-helpers/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 2.5.2
 
 - [fix] detect app.ts for runtime analyze
+- [chore] bump version of `esbuild`(0.14.14)
 
 ## 2.5.1
 
@@ -13,7 +14,6 @@
 - [chore] revert `@babel/plugin-transform-runtime` helpers value
 - [chore] bump version of `es-module-lexer`
 - [feat] support scan source code of import statement
-
 
 ## 2.4.3